refactor(/deputado-federal/[name]): extract expenses tab and cabinet data from page component
The ExpensesTab and CabinetTab components were created as a product of this extraction. The components that are only used by ExpensesTabComponent are now on the same directory of it. The new directory location for the ExpensesTabComponent and it's children components is: components/deputado-federal/[name]/tabs/expenses. The equivalent is true for CabinetTab component, wich is located in: components/deputado-federal/[name]/tabs/cabinet.